Transitioning to generic arkworks impl
