+++ src/fields/constants.rs
@@ -1,0 +1,71 @@
+/// (fp, u64)
+pub(crate) const FP_MODULUS: [u64; 6] = [
+    9586122913090633729,
+    1660523435060625408,
+    2230234197602682880,
+    1883307231910630287,
+    14284016967150029115,
+    121098312706494698,
+];
+
+pub(crate) const FP_MODULUS_MINUS_ONE_DIV_TWO: [u64; 6] = [
+    4793061456545316864,
+    830261717530312704,
+    10338489135656117248,
+    10165025652810090951,
+    7142008483575014557,
+    60549156353247349,
+];
+
+pub(crate) const FP_MODULUS_BIT_SIZE: u32 = 377;
+
+pub(crate) const FP_TRACE: [u64; 6] = [
+    8435453208297608227,
+    9853568280881552429,
+    7479357291536088013,
+    1657802422768920715,
+    16796279350917535980,
+    1720,
+];
+
+pub(crate) const FP_TRACE_MINUS_ONE_DIV_TWO: [u64; 6] = [
+    13441098641003579921,
+    14150156177295552022,
+    12963050682622819814,
+    828901211384460357,
+    8398139675458767990,
+    860,
+];
+
+pub(crate) const FP_QUADRATIC_NONRESIDUE_TO_TRACE: [u64; 6] = [
+    7563926049028936178,
+    2688164645460651601,
+    12112688591437172399,
+    3177973240564633687,
+    14764383749841851163,
+    52487407124055189,
+];
+
+pub(crate) const FP_ZERO: [u64; 6] = [0, 0, 0, 0, 0, 0];
+
+pub(crate) const FP_ONE: [u64; 6] = [
+    202099033278250856,
+    5854854902718660529,
+    11492539364873682930,
+    8885205928937022213,
+    5545221690922665192,
+    39800542322357402,
+];
+
+pub(crate) const FP_GENERATOR: [u64; 6] = [15, 0, 0, 0, 0, 0];
+
+pub(crate) const FP_TWO_ADIC_ROOT_OF_UNITY: [u64; 6] = [
+    9136220608200423841,
+    495125767030442947,
+    12992209043373160646,
+    4682582920187257812,
+    3526690454729915547,
+    15385663861525120,
+];
+
+pub(crate) const FP_TWO_ADICITY: u32 = 46;
