--- src/fields/u64/fr.rs
@@ -1,4 +1,0 @@
-// #[cfg(feature = "arkworks")]
-pub mod arkworks;
-pub mod fiat;
-pub mod wrapper;
